+++ src/main/java/likelion/festival/booth/ui/BoothController.java
@@ -5,14 +5,14 @@ import likelion.festival.booth.application.dto.BoothDto;
 import likelion.festival.booth.application.dto.BoothFilterDto;
 import likelion.festival.comment.appliction.dto.CommentRequestDto;
 import likelion.festival.comment.appliction.dto.CommentResponseDto;
-import likelion.festival.dto.*;
 import likelion.festival.booth.domain.Booth;
 import likelion.festival.booth.application.BoothService;
+import likelion.festival.like.application.dto.LikesResponseDto;
 import likelion.festival.menu.application.dto.MenuRequestDto;
 import likelion.festival.menu.application.dto.MenuResponseDto;
 import likelion.festival.comment.appliction.CommentService;
 import likelion.festival.service.ImageService;
-import likelion.festival.service.LikesService;
+import likelion.festival.like.application.LikesService;
 import likelion.festival.menu.application.MenuService;
 import lombok.RequiredArgsConstructor;
 import org.springframework.http.HttpStatus;
