+++ src/main/java/likelion/festival/like/application/LikesService.java
@@ -1,12 +1,12 @@
-package likelion.festival.service;
+package likelion.festival.like.application;
 
-import likelion.festival.dto.LikesResponseDto;
+import likelion.festival.like.application.dto.LikesResponseDto;
 import likelion.festival.booth.domain.Booth;
-import likelion.festival.entity.Likes;
+import likelion.festival.like.domain.Likes;
 import likelion.festival.exception.WrongBoothId;
 import likelion.festival.exception.WrongLikesKey;
 import likelion.festival.booth.domain.repository.BoothRepository;
-import likelion.festival.repository.LikesRepository;
+import likelion.festival.like.domain.repository.LikesRepository;
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Service;
