+++ src/main/java/likelion/festival/like/application/dto/LikesResponseDto.java
@@ -1,4 +1,4 @@
-package likelion.festival.dto;
+package likelion.festival.like.application.dto;
 
 import lombok.*;
